+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
-
+### Added
+- Alternative middleware that only checks but doesn't redirect if not logged
 
 ## [1.7.1] - 2019-03-23
 ### Fixed